Force PIP not to upgrade any Debian-installed Python packages.
This PR amends the `update-constraints.sh` script to automatically
extract from the filesystem the list of Python packages that have
already been installed by Debian, and that PIP should not attempt to
upgrade to any later version.

In addition to the normal `constraints.txt` file, it generates an
additional `pip-constraints.txt` file, to be used to force PIP not to do
anything with the system-managed packages.

This has the side-effect of ensuring that we never upgrade to any
version of setuptools more recent than 72 (known to cause issues with
some packages needed by the ODK), since the version provided by Debian
is 68.1.2.