Warn if system-managed Python packages differ from expected.
When building the ODK builder image, before attempting to install the
Python packages, we obtain the list of packages that have already been
installed by the system (i.e. by APT) and compare it with the
pre-computed list in `pip-constraints.txt`.

The two lists should normally always be the same. They MAY differ if,
between the time the constraints were last updated and the current
build, Ubuntu updated some of its Python packages (e.g. if they bump
`python3-setuptools` to a more recent minor version because of a
security vulnerability). If that happens, we allow the build to continue
(it may not necessarily cause a fatal error), but we log a warning so
that, should the build fail, we will know that we likely need to refresh
the constraints to reflect the updated Ubuntu packages.