Add new Makefile variable REASONER.
Instead of hard-coding `{{ project.reasoner }}` at several places in the
standard Makefile (whenever a reasoner is required), we add a new Make
variable REASONER which is set to the value of `{{ project.reasoner }}`,
and which is then used whenever a reasoner is required.

This allows:

(1) quickly changing the reasoner used by an entire project, without
having to change the ODK configuration and re-generate the Makefile
(e.g. for temporarily testing another reasoner);

(2) custom worklows in $(ONT).Makefile to automatically use the same
reasoner as the standard workflows, by using the REASONER variable
themselves instead of hard-coding a reasoner's name.

closes #1139