Create the JSON context file under src/ontology/config.
The _dynamic_files.jinja2 template is creating the context.json file
(list of prefixes used in the project) in src/config, but the standard
Makefile is expecting to find that file in src/ontology/config.

We amend the template to create the file at the location where the
Makefile expects to find it.

closes #1146