Do not stored derived file in src/ontology/config.
When building a SQLite product, we need to transform the JSON-formatted
src/ontology/config/context.json file into a CSV variant, for the
benefit of the semsql tool.

Since the CSV file is a derived file that people are not supposed to
edit (the source of truth is the JSON file), it is better to put that
file in the temporary directory rather than the config directory.