Do not depend on an inexisting variable.
If the project is not configured to use a context, then the target to
produce .db files does not need to depend on $(CONTEXT_FILE_CSV).

This still works because, without a context, the CONTEXT_FILE_CSV
variable is never declared, so it expands to nothing. But it is cleaner
to avoid relying on this.