Make sure SPARQL checks cover the entire ontology.
ROBOT's `verify` command does not follow OWL Import declarations, so the
checks it performs only applies strictly to the file it is given as
input.

Components and pattern-derived definitions are part of the "local"
ontology and should be covered by the SPARQL checks, and we ensure they
are by running the `verify` command on the `$(SRCMERGED)` intermediate
file, in which components and pattern-derived definitions (basically all
files listed in $(OTHER_SRC)`) are explicitly merged (but _not_ ODK
imports).

closes #1154