Move Ammonite to ODKFull.
Ammonite is no longer used by any standard, ODK-generated workflow -- it
was only used by the `make validate_idranges` workflow, which now uses a
dedicated program.

So we move Ammonite to the ODKFull image only, in line with the intended
usage of ODKLite -- ODKLite should contain everything that is required
by standard workflows, and nothing more.