Update some ODK programs.
Update some of the programs bundled with the ODK:

* Soufflé to version 2.4.1;
* KGCL-Java to version 0.5.1;
* SSSOM-Java to version 1.1.1.

Also add a note to remind us that we *cannot* update Jena to version 5,
as that version requires Java >= 17 and the ODK base image still
provides Java 11 for now.

There is a 3.0.0 version of Ammonite since september 2024, but we do not
upgrade it for now because (1) I don't know or use Ammonite enough (at
all, actually) to know whether the new major version is likely to break
any existing custom workflow that requires Ammonite, and (2) Jim Balhoff
is keen to migrate from Ammonite to Scala-CLI anyway.