Make sure /tools/.coursier-cache is user-writeable.
The ODKFull image defines the COURSIER_CACHE environment variable to
/tools/.coursier-cache, so Ammonite and Scala-CLI expect to be able to
store their downloaded dependencies in there.

If we run under a non-privileged user, we need to ensure that, if the
directory does not exist, it is created and given to the odkuser.

(If the directory already exists, it means a host directory has been
explicitly bound to it, in which case permissions will already be
correct, and we have nothing to do.)