Simplify Make conditionals.
Some pre-existing Make conditionals were using the following form:

  ifeq ($(strip $(VARIABLE),true))

This is because some VARIABLES were declared with a comment on the same
line, as in

  VARIABLE = true # some comment

The presence of the comment causes Make to treat the whitespaces on the
line as significant, meaning that VARIABLE in the example above has the
value " true ", instead of "true". Using the `strip` function in the
condition allowed to test against the intended value of the variable,
rather than the value effectively set.

But it is actually better to fix the declaration of the variable (moving
the comment to the line above), once and for all, than to strip
whitespaces whenever the variable is tested. This has been done for all
affected variables, so the calls to `strip` are no longer necessary.