Use Make conditionals for components.
Apply to targets of the components pipeline the same principle as for
mirrors and imports: bracket the entire section within a single Make
conditional on the COMP variable.

Also, use a Make conditional on the MIR variable to avoid re-generating
a component that is downloaded from a remote source when MIR is false.