Regroup all variables to control workflows.
Declare all the variables that control whether workflows are enabled
(MIR, IMP/IMP_LARGE, COMP, and PAT) in a single block near the beginning
of the Makefile.

This both ensures that all such variables are always declared before
they are used, and makes it easier for users to find them and understand
what they are for.