Replace SPARQL pre/postprocessing steps with pluggable ROBOT commands.
When preparing import modules, we do a few things:

(1) add a dc:source ontology annotation, derived from the version IRI of
the original ontology;
(2) remove all other ontology annotations, keeping only the newly added
dc:source;
(3) inject proper SubAnnotationPropertyOf axioms for properties
representing subsets and synonym types.

All those steps are currently performed by SPARQL queries. Here we
replace those queries by calls to the `odk:annotate` command, which
takes care of (1) and (2), and to the `odk:normalize` command, which
takes care of (3).

Of note, the fact that we are no longer going through a SPARQL
processing step means that we could end up with duplicated axioms with
different sets of annotations. Those were automatically merged as a
side-effect of the SPARQL processing (which involves dumping the output
of the SPARQL processing and re-parsing it again into OWLAPI objects).
Since we no longer benefit from that side-effect, we must explicitly
include a step in which we merge duplicated axioms (theoretically this
could be done with `robot repair --merge-axiom-annotations`, but
unfortunately this command does not behave exactly like we would [1]).

[1] https://github.com/ontodev/robot/issues/1239