Add a test for alignment with an upper ontology.
We add a new option in the 'robot_report' section called
'upper_ontology'. If set, it should be the (resolvable) IRI of an upper
ontology (such as http://purl.obolibrary.org/obo/cob.owl).

When set, a a new report is added to the list of ROBOT report, one that
tests whether all classes of the ontology are classified under one of
the classes of the upper ontology. The new report uses the same
parameters as the standard ROBOT reports regarding the file to perform
the check on ('report_on') and whether the check should be limited to
classes within the project's namespaces or not ('use_base_iris').

See #1175