Move the pattern_clean rule out of the PAT=true block.
Make sure the `pattern_clean` rule is always defined, even when PAT is
set to false. This is because that rule is explicitly called by the
general `clean` rule.

Of note, under PAT=false, `pattern_clean` can be called but it will not
clean _all_ the pattern-derived files, because the variables defining
the list of files to clean are themselves defined within the PAT block,
and moving them out of the block would be too much of a hassle for little
benefits.

While we are at it, this commit also removes an obsolete comment about
the prepending of `./` to prevent deleting files outside of the
repository, and clears some superfluous whitespaces.

closes #1176