Simplify the install_templates function.
Use list comprehensions to iterate over only the plain files first, then
over only the templates. This is no more efficient (we still need to
iterate over all the files twice), but is more readable.

When applying the templates, check whether the template is a "dynamic
pack" first, and only open the "derived_file" if it is not.

A further simplification would be to loop over all the files only once
and test for the template suffix once inside the loop:

    for f in files:
        if not f.endswith(TEMPLATE_SUFFIX):
            # Copy plain file
        else:
            # Apply template (single file or pack file)

but this would not allow to preserve the current behaviour of first
copying all the plain files before instantiating the templates (plain
files and templates would be processed in the order they are listed).