Allow to have a per-file installation policy.
Make the install_template_files function accept an optional `policies`
dictionary. A key is that dictionary is the pathname of a template file
(relative to the top level of a ODK repository); the value should be one
of:

* IF_MISSING: install the corresponding file into the target repository,
  only if it does not already exist;
* ALWAYS: always install the corresponding file;
* NEVER: never install the corresponding file;

The policy dictionary is queried for every single file to install
(whether it is a plain file or a file from a Jinja2 template), and the
file is installed or not based on the corresponding policy. If the
dictionary does not contain an entry for a file, the default policy is
IF_MISSING.

In the `seed` command, there is no policy dictionary and so all files
are installed (unless they happen to exist already, in the case the
`seed` command was called without the `--clean` option) and the command
behaves as before.

But the policy dictionary will make it possible to implement an `update`
command where we can selectively decide which files must be overwritten
when we update an existing repository.