Add a `update` command to odk.py.
Now that we can selectively install instantiated template files in a
directory depending on a per-file policy and on whether the target file
already exists or not, we can implement a repository update method
directly in the main odk.py script.

The command simply:

1. finds the '*-odk.yaml' configuration file, aborting if it can't find
   it (this also acts as a check that we are within a ODK repository);
2. loads the configuration;
3. declares which files must be ignored or overwritten;
4. instantiates the templates.

This is a much more straightforward process than the original
`update_repo.sh` script, which involved seeding a brand new repository
in a staging directory and then selectively copying files from the
staging directory into the real repository.

Also, because the update code is within the odk.py script, this means it
is within the ODK image. Therefore, when using that command, it is
always the latest available update code from the latest ODK version that
is immediately used. With the previous method, the `update_repo.sh`
script that was executed was the one that was installed with the
previous version of the ODK, leading to the recommendation of running
the script twice so that the second invocation would use the updated
code.