Remove the old update_repo.sh script.
That script will not be used anymore.

We keep the `update_repo` target in the Makefile, but make it invoke the
update command of the odk.py script. Using that target is not
recommended, it is kept for backwards compatibility only. (Its problem
is that the Makefile is overwritten during the update process.)