Install a helper update_repo script, update documentation.
Updating a repo can now be done by invoking `sh run.sh odk.py update`.

But for convenience, we install a dead simple script in
/tools/update_repo that merely call the above command, so that an update
can be triggered using `sh run.sh update_repo`.

This is because until now, most people never had to deal with the
`odk.py` script directly, since it was typically only used for seeding a
repository (which is only ever done once). It was also indirectly used
for updates (since an update involved seeding a brand new repository and
then copying its files to the current repository), but this fact was
hidden from the users.

So, I believe suddenly introducing the need to call that odk.py script
to update a repo could cause needless confusion in users' mind, so we
hide it behing a single `update_repo` command. This also has the benefit
of shielding us from any future potential change to the interface of the
odk.py update command.