Allow overwriting run.sh without causing a failure.
When updating, the shell executing the run.sh script (through which the
update command is executing, as any other ODK process) could fail
because the run.sh script is overwritten during the update.

To avoid that, we detect at the very beginning of the script if the user
is invoking the update command, and if it is, we make a copy of the
script in the tmp/ directory and immediately execute that copy, so that
any overwrite of the original script has no effect on the executing
shell.