Associate updating policies to shell-like globbing patterns.
Change the type of the policies argument to `must_install_file` from a
dictionary to a list of tuples, where the first item in each tuple is a
shell-like globbing pattern.

This is slightly less efficient as it requires iterating over the entire
list to find a matching policy (instead of the on-average constant
access time of a hashtable), but is more convenient as it allows to have
a single policy for an entire subtree. Notably, it allows to do

  policies.add(('src/sparql/*', ALWAYS))

to apply a ALWAYS policy to all files under `src/sparql`, without having
to enumerate them all.