Simplify the `prepare_initial_release` rule.
The `prepare_initial_release` rule does exactly the same thing as the
`prepare_release` rule, except that it also adds the release assets to
Git's staging area (so that they will be tracked by Git and later
committed).

Instead of duplicating the code of the `prepare_release` rule, we simply
make `prepare_initial_release` rule dependendent on `prepare_release`,
so that the only thing the rule specifically does is adding the release
assets to the staging area.