Do not attempt to commit if there is nothing to commit.
When we are not committing the release artefacts, it could happen that
`make prepare_release` does not modify any of the tracked files (e.g. if
there are no import modules that would normally be refreshed as part of
the release preparation), and in that case we should not attempt to
create a commit (otherwise Git would error out, complaining that there
is nothing to commit).