Let odk.py take care of committing the release artefacts.
Remove the `prepare_initial_release` target from the Makefile. Instead,
add a helper `show_release_assets` target that simply prints the list of
release assets; that target is used by the seeding command to know which
files should be forcefully added to the repository if the user chooses
to commit the release artefacts.

Replace the `commit_release_artefacts` project-level option by a
`--commit-artefacts` command-line option at the level of the `seed`
command.