Do not commit release artefacts by default (#1183)
We add a `-r`, `--commit-artefacts` option to the `seed` command.
When that option is enabled, we commit the release assets from the
very first build into the repository (assuming that `--skip-git` is _not_
used; under `--skip-git`, we are never committing anything no matter
what).

Adding the release assets to the repository is done directly from
within the `seed` command, which allows us to remove the
`prepare_initial_release` target from the standard Makefile (that
target was only ever needed by the `seed` command). But we do
add instead a small helper target, `show_release_assets`, that the
`seed` command is using to get the list of the assets to commit.

In .gitignore, we unconditionally ignore the top-level release artefacts,
so that they can never be committed inadvertently. If the user
chooses to commit release artefacts at commit time (using the
`--commit-artefacts` option), the `seed` command will forcefully add
the artefacts so that Git will start tracking them anyway.

closes #1144