Put the current ROBOT profile in the templates directory.
For now, when building the image, we download the ROBOT report profile
for the current version of ROBOT (the version used in the ODK) and put
it in /tools/robot_report_profile.txt. We do that so that we can check,
as part of the standard workflow, whether the custom profile that the
project may be using is not lacking any new tests that are available in
the current version of ROBOT (see #998).

Here, we simply move the location of the profile file within the image,
from

  /tools/robot_report_profile.txt

to

  /tools/templates/src/ontology/profile.txt

This does not change how the check mentioned above works (we just update
the path involved in that check), but this will have the benefit of
automatically installing the ROBOT profile when seeding a repository,
since the file is now within the template directory used by the seeding
system.

Therefore, we can remove the 'src/ontology/profile.txt' block from the
dynamic template pack (which needed to be manually updated by us
whenever we updated ROBOT -- something I don't think we have really been
doing).

But we do not want to _always_ install this profile.txt file in a seeded
repository -- we only need to do that when the use of a custom profile
is enabled in the configuration (that was the reason why the profile was
part of the dynamic template pack). To avoid installing the file when a
custom profile is _not_ enabled, we use the recently introduced
per-template policies system: we set the policy for the
'src/ontology/profile.txt' template to NEVER if a custom profile is not
enabled.

closes #999