Error out cleanly if the config file cannot be parsed.
Do not catch any exception thrown by the YAML parser in the
`load_config` method. Let the exception be propagated up to the level of
one of the Click commands and process it there.

"Processing" the exception includes:

(1) making a clean error message (including details about what the
problem is, if we got them from the YAML parser);
(2) raising a Click exception, which will in turn ensure that (a) the
error message is printed out, (b) the process is terminated, (c) the
process returns a non-zero exit code.

closes #1167