Remove obsolete documentation files.
The docs/ directory contains a lot of files that used to represent the
user documentation as published on
ontology-development-kit.readthedocs.io. That documentation has not been
updated for a while, and is now deprecated in favour of the OBOOK. The
contents of most if not all pages have been replaced by merely links to
the OBOOK, but even those links are not published anywhere nowadays.

So we remove everything.