Move some import-related logic to the odk.py script.
In order to make the Makefile template more readable, we need to move
some of the logic it contains into the odk.py script.

The general principle is that the configuration object that is given to
the templates should contain "pre-computed" fields that can be used
directly by the templates. This pre-computation will be done by the new
`_derive_fields()` method in the `ProductGroup` class and its derived
classes.

In this commit, we migrate some of the logic needed to deal with
imports. For example, instead of letting the Makefile template having to
deal with the possibility that the `module_type` of a given module might
not be set (in which case the template needs to get the value of the
`module_type` at the level of the entire ImportGroup), we ensure in the
`_derive_fields()` method of ImportGroup that all individual modules
have an explicit `module_type` (set from the group-level value if no
module type has been explicitly specified in the configuration file), so
that the Makefile template can simply use something like:

  {% if 'filter' == ont.module_type %}

instead of

  {% if ('filter' == ont.module_type) or
        (ont.module_type is none
         and 'filter' == project.import_group.module_type) %}

Likewise for the `slme_individuals` and `module_type_slme` parameters.

Also, we add a `special_products` derived field to the ImportGroup,
containing the list of all the modules that either (1) are marked as
"large", or (2) use a different type than the default, group-level
module type.

Overall, this makes the section of the template responsible for the
production of the rules for the individual modules easier to read.