Indent Jinja2 conditionals.
Indent Jinja2 conditionals (within the {% ... %} blocks to make the
template easier to read.

That is, instead of:

  {% if foo %}
  {% for b in bar %}
  {% if b.is_baz %}
  ...
  {% endif %}
  {% endfor %}
  {% endif %}

we do instead:

  {% if foo %}
  {%   for b in bar %}
  {%     if b.is_baz %}
  ...
  {%     endif %}
  {%   endfor %}
  {% endif %}

This does not solve *all* the readability issues of the Makefile
template, but at least it provides a relatively easy way to find, for
example, the `if` that corresponds to a given `endif`.