More formatting and comments.
Move the rule that produce the OBO version of an import module towards
the end of the imports section, so that it is not mixed with the rules
that produce the actual import modules.

Apply some minor Jinja2 reformatting, and also add a comment explaining
why we need non-generic rules for the "special" imports.

Include the type of the module in the generated Makefile comment, as
advanced users might find that useful.