Expand arguments list with one item per line.
When generating a ROBOT command that takes repeated arguments from a
list, write each item of the list on a separate line. That is:

  robot extract --base-iri IRI1 \
                --base-iri IRI2 \
                --base-iri IRI3 \
                ...

instead of:

  robot extract --base-iri IRI1 --base-iri IRI2 --base-iri IRI3 ...

as the second form could produce arbitrary long lines depending on how
many items we need to expand.

This makes the generated Makefile easier to read, at the expense of
making the *template* slightly harder to read. But since the Makefile is
what the users are exposed to, its readability is more important than
the readability of its template (which only ODK developers need worry
about).