Remove the 'fast_slme' module type.
Though never officially documented, 'fast_slme' was a module type that
was almost identical to 'slme', the only difference being that a "fast"
SLME module was produced without removing the original ontology
annotations and without adding a 'dc:source' annotation. Since those
tasks used to be performed by SPARQL queries, they were very
memory-intensive and could be a bottlenecks for large modules.

In fact, SLME modules that were marked as "large" were automatically
treated as "fast SLME" modules because of their high memory requirement.

Now that adding the 'dc:source' annotation is added by the ODK plugin in
a way that does not involve SPARQL, that step is as fast as it could be
and, more importantly, does not depend on the size of the import module
anymore. Therefore, there is no longer any real need for a "fast track"
for large SLME modules.

So we remove that fast track entirely. All SLME modules are now equal,
regardless of their size. For backwards compatibility, if a module is
explicitly configured as a 'fast_slme' module, we silently recognise it
as a standard 'slme' module.

This also means that for now, the only effect of the 'is_large'
parameter is to allow to selectively *not* refresh large import modules,
while refreshing the other, non-large modules.