Remove the *_terms_combined.txt files.
As suggested by a comment in the template, there is no real need for the
Makefile to catenate, sort, and "uniquify" the lists of terms we use as
a seed for imports (the automatically generated $(IMPORTSEED) and the
manually maintained $(IMPORT)_terms.txt). ROBOT is perfectly able to
accept:

* more than one --term-file;
* containing comments;
* containing unsorted terms;
* containing duplicated terms.

Anything we might do before passing the lists to ROBOT would in fact be
redundant with what ROBOT will do anyway, so we might as well let ROBOT
do everything.

This means that we no longer need any *_terms_combined.txt file.