Add an option to allow using an entirely manually curated import seed.
Add a 'scan_signature' option in the ImportGroup. If true (which is the
default), a "pre-seed" is extracted from the -edit file (and its
components, if any). That pre-seed is added to the main import seed
(tmp/seed.txt) along with any terms extracted from the DOSDP patterns
(if any). This is the behaviour of the current version of the ODK.

If that option is set to false, then the pre-seed is not extracted, and
terms from DOSDP patterns are not added to the main seed as well. This
means that the main seed can now only contain the custom seed optionally
set by the 'use_custom_import_module'. This, in turn, means that the
only terms that will effectively be imported will be those actively
selected by the user (either in the custom import module, or in the
individual *_terms.txt files).

If we do not scan the signature and there is no custom import module,
then the tmp/seed.txt file would effectively be empty, so in that case
we do not even generate the corresponding rule in the Makefile.

closes #638