odk.sh: Print message about memory settings on stderr.
The odk.sh wrapper scripts prints a message to let users know the
maximal amount of memory that is available to Java programs within he
container.

But the line emitted by that script prevents the `make docs` target to
work correctly, because the line ends up at the beginning of the
generated project-schema.json file.

So we print the message on stderr rather than stdout.