Fix stripping of annotation properties.
We add a new configuration option in the ImportGroup,
`strip_annotation_properties`. Setting it to False disables the default
behaviour (introduced in ODK 1.5) of stripping most annotation
properties from the import modules.

When enabled, the stripping of annotation properties is now also
performed on individual SLME modules (previously it was only done on the
merged import module and on individual "mirror" and "filter" modules).

closes #1093