Do not ignore module-level SLME parameters.
Declare the module-level `module_type_slme` and `slme_individuals`
options to be None by default.

This way, we can detect whether those options have been explicitly set
for a given individual module, and if so, use their values instead of
using the values from the group level.

Then, we can detect whether the SLME type of an individual module is
different from the group-level SLME type, and make sure we generate a
specific rule for that module in that case.

closes #1204