Add a 'merged' maintenance mode for SSSOM mapping sets.
Add a new value for the 'maintenance' field of a SSSOM mapping set
product: 'merged'.

A "merged" mapping set is obtained simply by merging other mapping sets
together. The sets that are to be merged should be listed in a new
'source_mappings' key; if that key is absent, then the set is made by
merging all other (non-merged) sets.