Add a 'custom' SSSOM mapping set type.
For consistency with other product types, add an explicit 'custom' type
of SSSOM mapping set product.

The generated rule for a 'custom' SSSOM set does nothing but emit an
error message, reminding the user that they must override the rule in
their custom Makefile.

(Users could also declare the set as being of type 'manual' -- which
creates a rule that does nothing else but touch'ing the file --, but the
'custom' type makes it more explicit that the set is to be generated by
some custom rule.)