Pre-compute mapping.source_file in ODK script.
When building a SSSOM mapping set by extracting mappings from an
ontology (maintenance type set to 'extract'), if the ontology to extract
the mappings from is not specified (no 'source_file' key), we default to
extracting from the preprocessed edit file.

We do not change that logic, but we move it from the Makefile template
to the odk.py script, so that the template can just dereference
'mapping.source_file' without having to check whether that field had
been set or not.