Allow using the SSSOM ROBOT plugin to extract mappings.
Add a new option to the SSSOM mapping set group: 'mapping_extractor'.
That option determines the tool to be used to extract mappings from an
ontology, to create the mapping sets of type 'extract'. The default
extractor is 'sssom-py', which corresponds to the existing behaviour.

The new value is 'robot', which causes the mappings to be extracted with
the sssom:xref-extract command of the SSSOM ROBOT plugin.