Do not display Scala-CLI version.
Scala-CLI takes **several seconds** to print a version message. Since we
are displaying tools' version infos whenever we run the `test` target
(because users might find useful to know which versions are actually
used), this means that most standard pipelines are paused for a few
seconds everytime just to get the version of Scala-CLI. This is
ridiculous.

Scala-CLI is not used in any standard pipeline anyway, so most users
would never care about which version is available in the ODK. Users who
do have Scala-based custom pipelines probably know enough about it to
check the version themselves if they so need.