Fix `make test PAT=false`.
The `dosdp_validation` target (invoked as part of `make test`) is
defined within the PAT=true conditional block, meaning the target does
not exist when PAT is set to false. As a result, trying to run the test
suite (or any other pipeline that depends on `test`, such as
`prepare_release`) with the pattern pipeline disabled can only fail.

We fix this by defining a dummy `dosdp_validation` target in the
alternative (PAT=false) block.