Rebuild components without using `make -B`.
Update all the `recreate-*` targets (used to forcefully rebuild
components) so that they do not rely on Make's `-B` option.

Since components can have diverse source files (templates, mapping sets,
and possibly anything in the case of custom-built components), there is
no single source file that we can `touch` in order to force a rebuild.

Therefore, we make every component depend on a "stamp" file in the
temporary directory, whose only purpose is to indicate that a component
has been built. Then, to force a component to be rebuilt, we instruct
Make to assume that the stamp file is "new", thereby automatically
trigerring a rebuild of the component.

Related to #1207