Forcefully refresh imports without using `make -B`.
Avoid using Make's `-B` option to forcefully refresh an import. Instead:

A) If refreshing all the imports (`refresh-imports` or
`refresh-imports-excluding-large`), we simply explicitly invoke the
`clean` target. It's a little bit "overkill", but if the user explicitly
wants to refresh ALL the imports, doing that from a clean state is
always a good idea anyway.

B) If refreshing all the imports but without mirroring
(`no-mirror-refresh-imports`), then we can't use `make clean`, because
mirrors are wiped out as part of the `clean` target. Instead, we trick
Make into thinking that the -edit file and all the *_terms.txt files
have been modified.

C) If refreshing a single import, we do as in (B), but only with the
*_terms.txt file of the desired import.