Allow updating the XML catalog.
Make it possible to update an existing catalog file when updating an ODK
repository.

This is done by:

(1) having all the ODK-managed entries (e.g. the ones for imports and
the like) in a dedicated group identified with 'odk-managed-catalog';

(2) having an update procedure that merges the template-derived catalog
with the existing catalog.

When merging, the 'odk-managed-group' in the existing catalog is
entirely replaced by the corresponding group from the template, whereas
all other entries are preserved unless they are already present in the
ODK-managed group.

closes #1219