Remove the Nodesource APT source.
Back when we first added Node.js to the ODK (as a dependency for some
features of the Ontology Access Kit), the version available in Ubuntu's
repositories was considered too old, so we added a third-party
repository maintained by Nodesource to get a more up-to-date Node.js
version.

Now however, the Ubuntu 24.04 repositories actually provide a _more
recent_ version than the one available in the Nodesource repository that
we were using, which means that, ever since we upgraded to Ubuntu 24.04,
we have always got Node.js from Ubuntu, and the Nodesource repository
has no longer any purpose.

So here we simply remove that third-party repository.