Automatically update import declarations.
When updating a repository, use the newly available `odk:import` command
from the ODK ROBOT plugin to automatically update the -edit file and
ensure it contains import declarations for any import module or
component declared in the ODK configuration.

This allows users to add an import or a component by simply declaring it
in the ODK configuration file then running `update_repo`, without having
to manually edit anything (both the XML catalog and the -edit file are
now automatically updated).

By default, the update process assumes that _all_ import declarations in
the -edit file are managed by the ODK (that is, they correspond to ODK
import modules or components). Therefore, the update process will also
_remove_ any import declaration that does not correspond to a
ODK-declared import or component (to handle the case where the user has
removed an import from the ODK configuration).

If, for any reason, an ontology has other "custom" import declarations
in addition to those corresponding to ODK-managed import modules or
components, the user can set the new project-level option
`preserve_non_odk_managed_imports` to True to avoid removing those
declarations.

closes #1225